Simplify the logic for checking button presses.
Previous versions used a counter to count button presses and then compare it with the internal value - which is logic already provided by the `button_X.was_pressed()` function. This allows us to greatly simplify the program and make it easier to understand as well, without needing to explain the complicated logic.